Membangun_model/requirements.txt
dagshub==0.6.9
mlflow==3.10.1
optuna==4.8.0
pandas==2.3.3
numpy==2.3.4
matplotlib==3.10.7
seaborn==0.13.2
scikit-learn==1.7.2
python-dotenv==1.2.1